% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasiSpectral.r
\name{quasiSpectral}
\alias{quasiSpectral}
\title{Compare mutant to wild type spectral count data, using "QuasiSeq" methodology}
\usage{
quasiSpectral(spectralDataAll, n.mut, n.wt)
}
\arguments{
\item{spectralDataAll}{data frame of gene/protein identifiers (geneName) and their spectral counts in mutant and wildtype groups.}

\item{n.mut}{number of mutant animals}

\item{n.wt}{number of wild type animals}
}
\value{
A dataframe with the log2 ratios, p-values, and q-values. Also eta0, the estimated proportion of null samples
}
\description{
Compare mutant to wild type spectral count data, using "QuasiSeq" methodology
}
